backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
supabase>=2.3.0
langchain-community>=0.0.16
faiss-cpu>=1.8.0
pypdf>=3.17.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
huggingface_hub>=0.20.0
requests>=2.31.0
